Adding hand-rolled convex hull factory

+++ src/Common/Worlds/Common.Worlds.Builder/Algorithms/ConvexHull/HullPoint.cs
@@ -1,6 +1,6 @@
 ﻿namespace Common.Worlds.Builder.Algorithms.ConvexHull;
 
-internal sealed record HullPoint( float X, float Y ) : IPoint {
+internal sealed record HullPoint( int X, int Y ) : IPoint {
 	public HullPoint(
 		IPoint point
 	): this( point.X, point.Y ) {
